modularized gui
