+++ MST-game-toio-30/Assets/Scripts/CatchTheFruit/CTFGUIManager.cs
@@ -9,36 +9,22 @@ using Cysharp.Threading.Tasks;
 public class CTFGUIManager : MonoBehaviour
 {
     static private NetworkManager m_NetworkManager;
-
     void Awake()
     {
-        m_NetworkManager = GetComponent<NetworkManager>();
+        var go = FindObjectOfType<GUIManager>();
+        m_NetworkManager = go.GetComponent<NetworkManager>();
     }
-
     void OnGUI()
     {
         
-        GUILayout.BeginArea(new Rect(10, 10, 300, 300));
-        if (!m_NetworkManager.IsClient && !m_NetworkManager.IsServer)
+        if (m_NetworkManager.IsClient || m_NetworkManager.IsServer)
         {
-            StartButtons();
+            GUILayout.BeginArea(new Rect(10, 10, 300, 300));
+            CTFStatusLabels();
+            GUILayout.EndArea();
         }
-        else
-        {
-            StatusLabels();
-        }
-
-        GUILayout.EndArea();
     }
-
-    static void StartButtons()
-    {
-        if (GUILayout.Button("Host")) m_NetworkManager.StartHost();
-        if (GUILayout.Button("Client")) m_NetworkManager.StartClient();
-        if (GUILayout.Button("Server")) m_NetworkManager.StartServer();
-    }
-
-    static void StatusLabels()
+    static void CTFStatusLabels()
     {
         GUIStyle style = new GUIStyle();
         style.fontSize = 40;
