+++ MST-game-toio-30/Assets/Scripts/GUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2fa4ea4ea5e1475e8970771354b9d4d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
